+++ spectest/call/call.0.go
@@ -12,7 +12,7 @@ type Module struct {
 	t0       []any
 	elements [][]any
 	memory   []byte
-	maxMem   int32
+	maxMem   int64
 	g0       int32
 }
 
@@ -366,7 +366,7 @@ func (m *Module) Xas_call_indirect_first() int32 {
 	_ = p0
 	{
 		t1 := m.f0()
-		t2 := m.t0[uint32(i32_const(0))].(func(v0 int32, v1 int32) int32)(t1, int32(2))
+		t2 := m.t0[uint(i32_const(0))].(func(v0 int32, v1 int32) int32)(t1, int32(2))
 		p0 = t2
 	}
 	return p0
@@ -376,7 +376,7 @@ func (m *Module) Xas_call_indirect_mid() int32 {
 	_ = p0
 	{
 		t1 := m.f0()
-		t2 := m.t0[uint32(i32_const(0))].(func(v0 int32, v1 int32) int32)(int32(2), t1)
+		t2 := m.t0[uint(i32_const(0))].(func(v0 int32, v1 int32) int32)(int32(2), t1)
 		p0 = t2
 	}
 	return p0
@@ -386,7 +386,7 @@ func (m *Module) Xas_call_indirect_last() int32 {
 	_ = p0
 	{
 		t1 := m.f0()
-		t2 := m.t0[uint32(t1)].(func(v0 int32, v1 int32) int32)(int32(1), int32(2))
+		t2 := m.t0[uint(t1)].(func(v0 int32, v1 int32) int32)(int32(1), int32(2))
 		p0 = t2
 	}
 	return p0
@@ -401,7 +401,7 @@ func (m *Module) Xas_store_last() {
 }
 func (m *Module) Xas_memory_grow_value() int32 {
 	t0 := m.f0()
-	t1 := memory_grow(&m.memory, t0, m.maxMem)
+	t1 := int32(memory_grow(&m.memory, int64(t0), m.maxMem))
 	return t1
 }
 func (m *Module) Xas_return_value() int32 {
@@ -564,10 +564,10 @@ func f64_const(x float64) float64 {
 	return x
 }
 
-func memory_grow(mem *[]byte, delta, max int32) int32 {
+func memory_grow(mem *[]byte, delta, max int64) int64 {
 	buf := *mem
 	len := len(buf)
-	old := int32(len >> 16)
+	old := int64(len) >> 16
 	if delta == 0 {
 		return old
 	}
