+++ spectest/global/global.0.go
@@ -8,7 +8,7 @@ type Module struct {
 	t0       []any
 	elements [][]any
 	memory   []byte
-	maxMem   int32
+	maxMem   int64
 	g0       int32
 	g1       float32
 	g2       float64
@@ -243,7 +243,7 @@ func (m *Module) Xas_call_indirect_first() int32 {
 	_ = p0
 	{
 		t1 := m.g4
-		t2 := m.t0[uint32(i32_const(0))].(func(v0 int32, v1 int32) int32)(t1, int32(2))
+		t2 := m.t0[uint(i32_const(0))].(func(v0 int32, v1 int32) int32)(t1, int32(2))
 		p0 = t2
 	}
 	return p0
@@ -253,7 +253,7 @@ func (m *Module) Xas_call_indirect_mid() int32 {
 	_ = p0
 	{
 		t1 := m.g4
-		t2 := m.t0[uint32(i32_const(0))].(func(v0 int32, v1 int32) int32)(int32(2), t1)
+		t2 := m.t0[uint(i32_const(0))].(func(v0 int32, v1 int32) int32)(int32(2), t1)
 		p0 = t2
 	}
 	return p0
@@ -263,7 +263,7 @@ func (m *Module) Xas_call_indirect_last() int32 {
 	_ = p0
 	{
 		t1 := m.g4
-		t2 := m.t0[uint32(t1)].(func(v0 int32, v1 int32) int32)(int32(2), i32_const(0))
+		t2 := m.t0[uint(t1)].(func(v0 int32, v1 int32) int32)(int32(2), i32_const(0))
 		p0 = t2
 	}
 	return p0
@@ -283,7 +283,7 @@ func (m *Module) Xas_load_operand() int32 {
 }
 func (m *Module) Xas_memory_grow_value() int32 {
 	t0 := m.g4
-	t1 := memory_grow(&m.memory, t0, m.maxMem)
+	t1 := int32(memory_grow(&m.memory, int64(t0), m.maxMem))
 	return t1
 }
 func (m *Module) f34(v0 int32) int32 {
@@ -360,10 +360,10 @@ func i32_const(x int32) int32 { return x }
 //go:nosplit
 func i64_const(x int64) int64 { return x }
 
-func memory_grow(mem *[]byte, delta, max int32) int32 {
+func memory_grow(mem *[]byte, delta, max int64) int64 {
 	buf := *mem
 	len := len(buf)
-	old := int32(len >> 16)
+	old := int64(len) >> 16
 	if delta == 0 {
 		return old
 	}
