+++ spectest/store/store.0.go
@@ -6,7 +6,7 @@ import "encoding/binary"
 
 type Module struct {
 	memory []byte
-	maxMem int32
+	maxMem int64
 }
 
 func New() *Module {
